LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/overlapping-navigations-and-traversals/tentative/cross-document-traversal-cross-document-traversal-expected.txt



Harness Error (TIMEOUT), message = null

FAIL cross-document traversals in opposite directions: the second is ignored assert_equals: first load event must be going back expected "?1" but got "?3"
TIMEOUT cross-document traversals in opposite directions, second traversal invalid at queuing time: the second is ignored Test timed out
NOTRUN cross-document traversals in the same (back) direction: the second is ignored
NOTRUN cross-document traversals in the same (forward) direction: the second is ignored
